pyproject.toml
[project]
name = "py-playwright"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pytest-playwright>=0.7.0",
]
